pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "repseq"
version = "0.1.0"
description = "Select representative bacterial isolates for long-read sequencing using phylogenetic + AMR diversity"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pandas>=2.0",
    "scipy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "biopython",
]

[project.scripts]
repseq = "repseq.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["repseq"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
